+++ tests/conftest.py
@@ -1,0 +1,4 @@
+pytest_plugins = [
+	"pytest_asyncio",
+	"pytest_homeassistant_custom_component",
+]